move functions to Sort.c
